// generated/src/app/api-models/matrix-determ-x.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class MatrixDetermX {
  matrixDeterminant: string;
  matrixSeq: number;
  xAxisDescription: string;
  ageFrom: number;
  ageThru: number;
  gender: string;
  salaryFrom: number;
  salaryThru: number;
  medicareStatus: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  rateType: string;
}